Fix failing test case when an anonymous function has the name "anonymous"
For more information, see
  - http://www.2ality.com/2015/09/function-names-es6.html
  - https://bugs.webkit.org/show_bug.cgi?id=7726
  - https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Function/name

For Node.js v6.0 (and more recent V8), functions created using `new Function`
are given the name "anonymous". MDN claims (at the time of this commit) that
the name should be an empty string, but the Webkit bug and the 2ality resource
claim that "anonymous" should be used. The referenced bug was in 2006, so may
no longer be applicable, but its presence in V8 suggests a change in thinking
possibly along a similar vein.