Resolve Node v0.10 issue
In Node v0.10, `Buffer.byteLength` requires a `string`. In later
versions, the function supports providing a `Buffer` object. The
current workaround is just to convert the response content to a
string when determining the content length. A better workaround
would be to sniff the Node version and only convert to a string if
necessary. Added TODO items in the source to this end.