Remove unreachable special cases.
The case where the first argument passed to ibeta_derivative is zero cannot
occur because d1, d2 and x are all positive (for the
latter, notice that for x <= 0, ibeta_derivative will never get called). Hence,
d2/(d2 + d1*x) and (d1*x)/(d2 + d1*x) can't be equal to zero, as this would
imply that either d2=0 or d1*x=0. From a similar reasoning, it follows that
the first argument cannot be one.