Set the clone depth to 10
If the depth is too shallow, build failures may occur due to an
existing build attempting to shallow clone a repo and checkout a
particular commit. If the depth is less than the number of commits,
the build will error as the cloned repo does not have the commit
reference in its tree. Hence, we want a commit depth which is not
large (otherwise, slow clone times), but not too small that the
pace of development outpaces the build turnaround time.