Rename tools and commands
The reason for renaming is to reflect the difference between
a "module" and a "package". A module is a single file which exports
one or more values. A package is a collection of one or more
modules intended to be used for a related purpose. Hence, as the
tools list "packages" rather than modules, renaming seemed
appropriate.