Retry if goodness-of-fit test fails the first time
The goodness-of-fit tests, as they stand now, are probabilistic.
On rare occasion, these tests fail because the null hypothesis is
rejected more than an allowed threshold. An occasional failure is
acceptable, as such failures should be expected. In most cases, we
can trust that the next goodness-of-fit test will pass. This commit
addresses this by rerunning the tests if they fail the first time.