Filter folder names only in child directories
This resolves an issue encountered on Travis CI, where the repo
is cloned to a '**/build/*' directory. Path filters for the `find`
command were filtering for a 'build' directory anywhere along a
file path. Thus, no tests were found on Travis CI during the build
process. By restricting filtering to child directories, we address
this CI bug, as well as the possibility of other development
environments having parent directories with filtered folder names.