Update date string format
Older browsers vary in date string formats which are parsed. This
change adds time info, matching the Date.toString() format. For a
guide on date string parsing and browser support, see

http://dygraphs.com/date-formats.html
http://blog.dygraphs.com/2012/03/javascript-and-dates-what-mess.html

Needless to say, the date string format used in this commit will
NOT be supported in old (old) browsers, including IE8. If IE8
support is necessary, will need to detect and then use alternative
tests.