Resolve include directories as relative paths
This change is necessary to support building with GYP on Windows.
On AppVeyor, we run GYP via MSYS2 (Cygwin). When specifying that
include directories be resolved as absolute paths, AppVeyor would
error, complaining that headers files could not be found. Resolving
include directories as relative paths "fixed" the problem.

For an example build, see

https://ci.appveyor.com/project/kgryte/stdlib/build/2060/job/3lcwn1lxdi97wlr0