Use POSIX path convention
Requiring POSIX paths was required to successfully build on AppVeyor.
When allowing Node.js to return paths according to the host platform,
builds would error as files could not be located. This is possibly
due to running GYP via MSYS2 (Cygwin), where paths are being
converted between Windows and POSIX path conventions. By specifying
a single convention, we are able to "fix" the issue.

For an example failed build, see

https://ci.appveyor.com/project/kgryte/stdlib/build/2060/job/3lcwn1lxdi97wlr0