Handle small alpha parameters correctly
For alpha values smaller than one, the algorithm by Marsaglia & Tsang is not applicable. This commit implements the strategy suggested by the authors to handle this special case: To generate a random variate from a Gamma distribution with parameter alpha, we can alternatively draw from `Gamma( alpha + 1, beta )` and multiply the result with a uniform random variate raised to the power `1/alpha`.