Refactor to prevent NaN return values
This is not a regression, although the tests are not in the range of one machine epsilon to Julia anymore. The formula for the PDF used before was the same as Julia's; it had the deficiency to return `NaN` for perfectly valid input arguments as in the call `pdf( -11.923, 7.12, 0.013 )`. Calls such as these should have returned `0.0` instead of `NaN`, which they now do.