Guard against case where NaNs can occur
Code now addresses the case where c = 0 and the gamma function call evaluates to infinity. The result of the multiplication 0 times Inf is not defined, hence NaN was returned. However, the gamma function call should actually return a finite number, one that is too large to be represented in JavaScript. Hence, the summand should evaluate to zero in this case.