Remove support for providing an `equalDims` option
The `equalDims` option activated a path which used function
generation. Upon performance testing, this path did not lead to
hypothesized performance benefits. Instead, performance was
significantly slower than using the recursion code path. A reason
for the decreased performance is that function generation required
evaluating a string and subsequently compiling that string into a
function each time the function was called. If function generation
was faster, perhaps we'd see increased performance. But as it is,
the overhead is not justified.