Move test value generation to benchmarked expression
Previously, the test value was generated during the "setup" phase.
This resulted in benchmarks which were not oranges to oranges
comparable. For other languages, we generate new test values for
each iteration. Hence, to ensure similar tests, we needed to move
test value generation to be part of the benchmark.