Print an error message rather than hard fail
This commit changes installation behavior for Node.js native add-
ons. Instead of loudly failing (exiting with code 1), the commit
modifies the install recipe to print an error message, exit with
code 0, and attempt to install the next add-on. This change relies
on the assumption that failure to install an add-on is not fatal.
In other words, packages have a mechanism for coping with failed
add-on compilation; e.g., they have a pure JavaScript fallback.