Remove C API compat
This commit removes a parameter which allowed compiler optimizations
due to hardware efficiencies. Notably, older platforms could perform
faster comparisons with integers than doubles. Thus, in order to
flag a particular code path, the C implementation on which this
function is based included an additional integer parameter which
indicated if a double was zero in order to avoid checking if the
double was zero directly. Such strategies are not necessary in
modern environments where double comparisons are on par with
comparable integer comparisons.