Refactor by moving private methods to prototype
This commit does introduce a performance regression (according to
local benchmarks). This is most likely due to a) having to walk
the prototype chain to access the private method and b) the need
to optimize away property and array access. This stated, this
commit does result in a performance boost for instantiating an
ndarray, as dynamic code evaluation is no longer necessary during
instantiation.