Disable running examples and add FIXMEs
Emscripten glue code appears to not support *only* providing a
`buffer` in order to define external memory. When a `buffer` is
provided, the glue code skips setting the `wasmMemory` property
of the `Module` object. When the module goes to instantiate a new
instance, it checks that the `wasmMemory` value is a `WebAssembly.
Memory` instance, but, because the property has not been set, the
module errors. This behavior does not seem correct.

For code references, see

-   https://github.com/kripken/emscripten/blob/655cec1975e784985d25bc80c5438d715ac5ec13/src/preamble.js#L2155
-   https://github.com/kripken/emscripten/blob/655cec1975e784985d25bc80c5438d715ac5ec13/src/preamble.js#L1141