Remove `async` code block configuration setting
This change ensures consistent behavior for synchronous and
asynchronous code blocks. Previously, if a synchronous code block
intentionally threw an exception, nothing was printed to `stderr`,
which was not consistent with code blocks which intentionally
threw an exception asynchronously. Additionally, setting an `async`
option made things more complex than they needed to be.