Set the `--build` flag when installing `emsdk`
This commit sets the `--build` flag to `Release`, as, by default,
`emsdk` builds with additional debug info which significantly
inflates the final build size (~10GB). Setting to `Release` ensures
that the debug info is not included, reducing the size of `clang`
to around `~1.7GB` on my Macbook. While not "small", this does
significantly reduce the footprint.

References:

-   https://github.com/juj/emsdk/blob/2324e5d8f419483bd4025c6ebb4eef1635d35625/emsdk#L1996
-   https://github.com/juj/emsdk/issues/109