Update namespace
This commit reverts making the plot "namespace" something of a
hybrid, where it is both non-conventional (i.e., a function, not
an object) and conventional (i.e., has sub-namespace members).
This led to confusion when generating a tree bundle where we need
a more strict delineation to determine how to interpret. As it
was, when generating a tree bundle, the hybrid nature of the
namespace led to attempting to overwrite a read-only property,
thus causing the bundle to error and, effectively, be unusable.