Explicitly set environment variable
For whatever reason, on Windows, the package name is not resolved
by plucking the project name from the root `package.json`. Here,
we resort to just hardcoding the project name. We may need to
revisit this issue in the future if we need to test npm installation
on Windows machines outside of CI.