Sniff umask behavior
This commit attempts to detect "anomalous" process.umask behavior.
In certain environments on Windows, process.umask does not exhibit
expected behavior; e.g., failing to set a new mask. Here, we add a
script to try and detect deviant behavior and skip tests if this
behavior is detected. This allows us to avoid failing builds on,
e.g., AppVeyor.

- https://ci.appveyor.com/project/kgryte/stdlib/build/job/7o6qgf25sovv9bea