Simplify implementation to ignore negative zero
This commit was done due to the fact that only `-0.0 - 0.0 = -0.0`,
which is not possible if `-0.0` and `0.0` are correctly assigned to
the minimum and maximum values, respectively. Accordingly, the sign
of zero is not relevant when computing the range. We can simply
normalize +-0 to +0 and achieve the same result. This commit uses
such normalization.