Fix references to function names in parent scope
This commit addresses bugs introduce after minification (and
mangling). Because the function references occur in evaluated
strings, several bugs can be encountered. For example, the
minification engine may decide to eliminate what it considers
unused code/variables. Next, a minification engine may rename the
variables, thus introducing a discrepancy in the evaluated string
which relies on the source variable names. The workaround used in
this commit is to bind the functions directly to the evaluated
function, where property names are less likely to be mangled.