Refactor to delay performance cliff
This commit seeks to avoid the performance penalty incurred when
resolving a value's constructor name. In short, resolution of
typed array constructor names was found to be significantly slower
than, e.g., generic arrays. This performance slowdown was affecting
downstream consumers. While possible that the performance profile
is specific to V8, such is the world we live in, given our reliance
on Node.js. The basic thrust of this refactor is to rely on
instanceof checks which are suitable for brand checking, as long
as a value is from the same realm. For cross-realm objects, we are
still able to resolve them, albeit more slowly.