Increase tolerances
Recent changes introduced by changing the default PRNG to MT19937
resulted in different values being provided as inputs. For one of
the input value combinations, the delta was significantly higher
than any combination provided as a result of invoking the previous
generator. As a result, in order for the tests to pass, we need
to increase the tolerance.