Fix assertion
This commit replaces a higher level assertion package with a lower
level variant, as the assertion is used in conjunction with another
assertion which overlaps in functionality (a) and (b) the assertion,
as used, checked for both primitive and object variants. In short,
this commit should introduce a marginal performance improvement
when sanity checking provided parameters.