Lowercase property name
This change stems from changes to how PRNG state is stored along
with an associated seed. As a consequence, whenever a PRNG state
is set, so, too, is the PRNG seed to ensure consistency and
reproducibility. In which case, a PRNG seed is no longer a "constant".
Hence, we lowercase the property name to reflect this fact.