Restrict benchmarks to float64
Upon attempting to run benchmarks for multiple array data types,
encountered a perf cliff when running on Node.js. Encountered a
similar perf cliff for `@stdlib/strided/math/abs`. This is likely
due to polymorphism causing a de-optimization in V8.

We restrict to `float64` here for convenience and to allow comparison
to, e.g., `dmap`.