Revert enumerability change due to performance impact
Setting inherited "private" properties to non-enumerable has a non-
negligible performance impact. While possible for a user to extract
all enumerable keys from an ndarray, this should not be common. For
the stdlib REPL, we should handle ndarrays as a special case when
displaying as a return value. This applies similarly to other JS
objects, such as typed arrays.