Add workaround for browserify not honoring `process` built-in
Browserify seems to expect that `process` will be used as a global,
but not a Node.js module which is required. Accordingly, as we use
`require( 'process' )` in the project, whenever we bundle for, say,
our custom ESLint rules, browserify complains about not being able
to walk our dependency graph while looking for a `process` module.
In which case, the workaround in this commit is to simply toggle a
flag which tells browserify to ignore missing modules. This is a
potential footgun, but hopefully missing packages/modules will be
caught before initializing the dev environment.