Allocate output array within function scope
Allocation and subsequent reusage from outside the calling scope
(e.g., in the parent scope) seems to trigger a de-optimization in
V8 (Node.js version 14 during profiling). This would seem to
indicate that thread safety is a factor in whether V8 is able to
optimize a particular function. Moving output array allocation
into the calling function's scope eliminated the de-optimization.