Reorder includes and fix buffer index resolution in C
This commit fixes a bug stemming from a failure to account for the
fact that a buffer index in JavaScript is not the same as a buffer
index in C. In the former, we are indexing into a typed array, so
a view index is the same kind of index as when indexing into the
underlying data buffer directly. However, in the latter, the buffer
index needs to be computed based on the byte offset. We're not
indexing into a typed array, but a byte array, so our index needs
to be translated into a byte index.