Refactor to use specified byte widths
This work is motivated by wanting to call into C from JavaScript
and have fixed widths for ndarray properties, such as index modes,
dtype, and order. Enum constants must fit in an int, but compilers
are free to use the minimum dtype necessary to hold the constant
value, and the dtype may vary within the same enum. Accordingly,
in order to ensure fixed byte widths, we needed to specify such
widths.