Specify workflow concurrency groups
This ensures that only the latest action for each workflow runs.
If multiple commits happen in quick succession (or at least before
a workflow run finishes), we are likely to only be interested in
the results for the last commit. This does, however, complicate
identifying which commit led to a failed build, but so long as logging,
etc is plentiful, we should be able to track down the issue.

At a minimum, specifying concurrency groups is necessary to reduce
our carbon footprint.