Use heuristic to test for a constants package
This heuristic is imperfect, as it relies on the assumption that
there are no sub-sub-namespaces containing constants.

This heuristic refactors changes introduced in

https://github.com/stdlib-js/stdlib/commit/021f7a3a050d2189c740f5b3cb714aab283b4990#diff-84df3a2cc3024abd4ffe88455d078865fa5f6ff80855a0bc8891f9b69d73e777

The previous implementation was flawed and problematic. Namely,
each time we loaded this lint rule, we needed to resolve the list
of namespaces, which meant reading the file system. After profiling
our linting, essentially 90% of the time was spent globbing the file
system.

We did not notice a slowdown immediately after introducing the
previous implementation because the lint rule ever worked as intended.
The reason is this line:

https://github.com/stdlib-js/stdlib/commit/86ffd643eb397ffeacf7b55a97108e256b742706#diff-13e457cd70693c0e55e32c45f[…]24d4ad6b61a212728da3560427edL28.

When we run linting, we don’t actually run the files from `@stdlib/_tools/*`.
Instead, we run from an ESLint plugin bundle placed in `node_modules`
upon `make init`.

Accordingly, in the line above, we never actually resolved the project
root directory. We were actually resolving a non-existent path in our
ESLint plugin directory.

Accordingly, the ESLint rule never worked as intended, as we were never
resolving a list of stdlib namespaces. The namespace array was empty.

So when I made the change in the above file on Aug 20, we started
walking up the file tree until finding the root directory. From there,
we resolved a list of namespaces. And, as might be expected, resolving
the list of namespaces takes a while, as we need to first resolve all
packages and then filter for namespaces. And given how we do linting,
we were having to resolve a list of namespaces for each file being linted,
leading to a significant slowdown.

TLDR; the change introduced in this commit drops the time needed to
lint a JavaScript file from upward of a couple of minutes to a few
seconds.