Adjust rule config to warn rather than error
Having this rule set to `warn` was problematic, as the current
implementation flags even empty directories as needing to be exported
to a namespace.

We could adjust the rule implementation to also check for `package.json`
files in the candidate directories. However, this is still problematic,
as, for legitimate reasons, a local package may not yet be added to
the Git index (i.e., tracked) and thus not be part of the public
codebase.

This lint rule could be configured to `error` in a different context
such as a CI cron job which ensures that all namespaces in the public
repository satisfy the lint rule requirements.