Refactor to require explicitly specifying array data types
This allows support for complex number data types which build atop
built-in floating-point data types. Without specifying the dtypes,
we don't have way to resolve whether a Float64Array should be
interpreted as a Float64Array or as a Complex128Array. We'd need
to provide additional information. Explicitly specifying the dtype
is that information, as it informs the C layer how it should
interpret a provided typed array.