Begin search based on the file's directory
This allows support for, e.g., editor linting outside of the stdlib
project. By not specifying the `basedir`, `resolve` will begin
searching from the top-level `node_modules` directory. We did not
observe failures previously as we explicitly set the NODE_PATH env
var during linting. With this change, setting the NODE_PATH env var
should no longer be necessary in order for this lint rule to be
used within the project.