Rename variable
This stems from investigating https://github.com/stdlib-js/math-base-tools-evalpoly/actions/runs/3062614392/jobs/4943787043.

In short, Istanbul stores coverage data on the global `this`
context. To access the global `this` context, Istanbul used the
Function constructor: (Function(){'return this;'}()). Accordingly,
when a module imports `@stdlib/function/ctor` and assigns to
`Function`, this overwrites the global variable. Due to variable
hoisting, the `Function` declaration was been hoisted to the top
of the instrumented file's scope. It's initial value being undefined,
as assignment took place later in the instrumented code. However,
Istanbul's attempt to access the global `this` happens as the first
line; hence, the issue. Istanbul was attempting to invoke the built-in
`Function` constructor, but was unable to do so due to us redeclaring
it via an import.

This commit simply changes the variable to avoid this global
reassignment.