Remove automatic installation of native add-ons upon project install
Installing native add-ons can be done in a separate step by running
make install-node-addons. Currently, this is a sticking point for
new devs who may conclude that something is wrong with their
installation if all native add-ons fail to compile. In most instances,
devs don't need to have native add-ons installed.

Furthermore, it probably is not necessary to re-install native add-ons
simply because a dev dep was added/removed from the project. We should
really only need to recompile upon switching Node.js versions. And
with N-API, we should only need to recompile if we're hoping to use
native add-ons on a very old Node.js version. In short, in general,
we shouldn't need to compile/recompile. As such, removing this step
from the general project install seems reasonable.