Fix variable name bug
Previously, if a macro is used more than once within the same block,
one would encounter redefinition errors. This commit addresses this
bug by creating a unique error variable name based on the output
variable name. This assumes that, if a macro is used multiple times
within the same scope, the output variable name is unique for each
macro invocation.