Address invalid date objects upon parsing response headers
Ever since GitHub migrated to its calendar versioning scheme, we've
been getting periodic errors when attempting to log rate limit info.
E.g.,

https://github.com/stdlib-js/stdlib/actions/runs/3810868110/jobs/6483195715

This seems to be due to missing header info. The workaround in this
commit is to default to an hour from now, according to machine local
time.

For GitHub docs which state that the rate limit headers should be
present, see https://docs.github.com/en/rest/overview/resources-in-the-rest-api?apiVersion=2022-11-28#rate-limiting.