Update description
This addresses the concerns raised in https://github.com/stdlib-js/stdlib/issues/837.
Namely, the value returned by `os.arch` is the CPU architecture for
which the Node.js binary was compiled, which may **not** be the same
as the actual host CPU architecture (e.g., if a binary is run on
a different architecture).

This commit aims to address the ambiguity of the previous description.