feat!: convert `Collection` to generic
This commit converts `Collection` to a generic in order to better
preserve type information. One example is when returning an element
from a collection and wanting to preserve the returned element's
type information. This change depends on default types, support
for which was added in TypeScript version 2.3, which is beyond the
current project support version of 2.0. Accordingly, this commit
increases the minimum TypeScript version and sets that version to
4.1, which should include most new TypeScript features that we
want to use in TypeScript declarations moving forward.

BREAKING CHANGE: `Collection` is now a generic

To migrate, users should upgrade their TypeScript version and
provide element type information when using the `Collection` type
(e.g., `Collection<number>`). By default, the element type is `any`
which is equivalent to previous behavior.