feat!: update callback signature
This commit reduces the number of provided arguments to the provided
callback function from 5 to 4, by consolidating the arrays into a
single argument. This brings this package in line with, e.g.,
`@stdlib/strided/base/unary-by` and others, where we seek to minimize
the number of arguments and also ensure consistency with future API
additions (e.g., `map3`, etc).

BREAKING CHANGE: consolidate array arguments into a single argument

To migrate, users should access the input arrays via the fourth
argument, where the first element is the first input array and the
second element is the second input array.