feat!: consolidate callback arguments
This commit updates the invocation signature of the provided callback
by providing only 3 arguments, rather than 6. By consolidating the
array element indices as a single argument, we make the behavior
more consistent with other packages in stdlib. For example, see the
array/base/* packages.

BREAKING CHANGE: provide index arguments as an array

To migrate, users should access the indices for the various dimensions
by accessing the respective elements in an index array argument, which
is provided to the callback function upon invocation.