fix: flatten nested arrays in (co)lexicographic order
This commit addresses a bug in flattening behavior. Prior to this
commit, if provided a nested array, the implementation would flatten
in lexicographic order (i.e., innermost nested array dimensions
iterate fastest). However, this lead to surprising behavior when the
desired storage order was column-major in which an ndarray did not
match the visual layout of the input nested arrays. This commit
fixes this issue and matches NumPy behavior in which nested arrays
are flattened such that the storage order matches the desired output
array storage order.

Ref: https://github.com/stdlib-js/stdlib/issues/268