build: change trigger to ensure availability of environment vars
This PR changes the `pull_request` trigger of the workflow to
`pull_request_target` to ensure that the environment variables
are available in the workflow when run from forks. The permissions of
the workflow are limited and we do not run workflows for untrusted pull
requests, mitigating the security risk of this change.