build: update scripts and move config to separate file
This commit moves the configuration for `discrete-uniform` as this
PRNG is unique in that it requires the `prng` option to be a PRNG
which generates integers, not uniformly distributed numbers on the
interval [0,1). The scaffold is not currently equipped to handle
this special case. Thus, in order to avoid clobbering the existing
package in the future, the configuration for this package is moved
to a separate file mainly for purposes of documentation, rather
than actual usage. If the file is used, manual edits will likely
need to be made to the updated package.