feat!: add `factory` method and remove options support
BREAKING CHANGE: use `factory` method for seeding PRNG

To migrate, users should use the `factory` method to create a
seeded PRNG. Previously, users could provide options directly to
the main export and `ndarray` method. This capability has been
replaced with a `factory` method. This change aligns with
conventions found elsewhere in the project where the main export
uses an unseeded PRNG, and, in order to create a seeded variant,
one needs to use a factory function.