feat!: return a `complex64` dtype ndarray if provided a Complex64 scalar
BREAKING CHANGE: return same dtype when provided a complex number instance

To migrate, users relying on the previous behavior where anything
complex-like resulted in a `complex128` ndarray should explicitly
set the output dtype to 'complex128'. By default, the function will
return an ndarray having the same dtype as a provided complex number
instance. When a user wants to explicitly upcast a Complex64 scalar,
the user can explicitly set the `dtype` option. The previous default
behavior is undesirable, as we disregarded the explicit type info
of a provided complex number scalar.