fix: avoid incorrectly prepending code blocks in sections preceding "Usage"
The previous code made assumptions regarding how README files are
organized. Notably, it assumed that the first code block in a
document contained a single line of code defining the `require`
statement for a corresponding package. For the most part, this
assumption holds true; however, it is not always true. E.g., if
one wants to include code blocks in the introduction of a README,
the previous code would incorrectly prepend that code block, rather
than the first code block in the "Usage" section, leading to subsequent
lint and evaluation errors. While the changes introduced in this
commit are not fully robust, as we still assume that the first code
block in the "Usage" section is the `require` statement and further
assumes that the code block is immediately preceded by a Markdown
heading having the value "Usage", it should be more robust than the
previous code based on project conventions.