feat!: rename exported aliases
This commit renames aliases to more naturally map to package names.
Previously, exported aliases attempted to match the global REPL
namespace aliases. This made more sense when the main distributable
artifacts were a "flat" and "tree" bundle and we wanted to have
similar naming conventions across both. Times have changed, however,
and the use of REPL aliases in the "ndarray" namespace feels awkward
and are hard to predict from a user POV, without knowing the
conventions of the REPL global namespace. Accordingly, this commit
changes aliases to what we'd typically use for a nested namespace.

BREAKING CHANGE: rename exported aliases

To migrate, users should consult the documentation for the updated
aliases. For the most part, updating code should be as simple as
dropping an `ndarray` or `nd` prefix. For others, they should more
directly match, or be inferrable from, the package name.