fix: replace use of type guards with boolean return type
The type narrowing behavior of type guards makes only sense when the
runtime checks match the type guards. For example `value is string`
should be used with `typeof value === 'string'` and not with a function
that checks for whether a value is a snakecase string. In the latter
case, the type guard would incorrectly narrow the type of the value to
`never` for a string that is not snakecase.