feat: Improve example in index.js for constants/uint8 namespace  This commit enhances the example code in index.js within the @stdlib/constants/uint8/examples directory. The updated code now showcases the available uint8 constants by listing their names in the console output.
This update aligns with RFC #1565, which aims to improve the README examples
of the constants/uint8 namespace package. By enhancing the example code in
index.js, users can better understand how to access and utilize the uint8
constants provided by the package.

The revised code imports the constants module from @stdlib/constants/uint8,
extracts the names of the available constants using Object.keys(), and then
prints out these names to the console. This provides a more informative output
that demonstrates the functionality of the constants/uint8 namespace.



Signed-off-by: Patel Prince <151719092+prince-patel23@users.noreply.github.com>