feat!: remove `order` argument from `ndarray` method
This commit removes the `order` argument, as this argument is not
necessary when providing multiple strides. The layout is conveyed
by the strides, not the order string, and requiring that both
strides and an order string be provided allows for the potential
for conflicting information (e.g., an order string of "row-major",
while the strides are in "column-major" order due to taking the
transpose, etc). In short, the strides are the source of truth.

BREAKING CHANGE: remove `order` argument from `ndarray` method

To migrate, users should drop the `order` argument when invoking
the `ndarray` method.