build: do not generate providence metadata
When generating provenance attestations within GitHub Actions workflows,
it is required that the repository in which the publish workflow is
executed matches the repository specified in the metadata. This assumption
does not hold with our decomposition of the monolithic repository into
multiple repositories. As such, we disable the generation of provenance
metadata.