feat!: add support for stacks in `blas/sdot`
BREAKING CHANGE: return an ndarray, rather than a scalar

This commit changes the return value from a scalar to an ndarray. Previously, the function only operated on one-dimensional ndarrays and returned a scalar value. Now, the function operates on ndarrays of arbitrary rank and always returns an ndarray. In order to migrate, for one-dimensional input ndarrays, users should call `out.get()` in order to retrieve the scalar dot product.

PR-URL: https://github.com/stdlib-js/stdlib/pull/2895
Co-authored-by: Athan Reines <kgryte@gmail.com>
Reviewed-by: Athan Reines <kgryte@gmail.com>