feat (stats/incr/nanwmean): adds nanwmean package to the stats/incr/* namespace
This commit adds a package which provides a way to compute the mean of a set of numbers while ignoring NaN values.
It is made to address [RFC] Issue #5628, and as suggested in the issue, it is based on a thin wrapper around wmean, similar
to the relationship between nansum and sum, mainting API consistency and design. This commit includes appropriate documentation and tests
for the new purpose of the package, styles of which are consistent to the stats/incr/* namespace.

Fixes: #5628 [RFC]
Private-ref: https://github.com/stdlib-js/stdlib/issues/5628
Authored-by: Don Chacko <donisepic30@gmail.com>