feat(stats/incr/nanmcovariance): add nanmcovariance to the stats/incr/* namespace
This commit adds the `nanmcovariance` module to the `stats/incr/*` namespace, providing a way to compute a moving unbiased sample covariance incrementally,
while handling NaN values. This commit was made to address Issue #5567 and as suggested in the issue, it is based on a thin wrapper around wmean, similar
to the relationship between nansum and sum, mainting API consistency and design. This commit includes appropriate documentation and tests
for the new purpose of the package, styles of which are consistent to the stats/incr/* namespace.

Fixes: #5567 [RFC]
Private-ref: #5628
Authored-by: Don Chacko <donisepic30@gmail.com>